Renamed Java directory to JavaFX.
Added a Java Console project for testing and experimenting.

Started Debugging LibOgame login system with the help of the JavaConsole Project.

I had to change the LoginActivity Class of DMB for Android since the libOgame API changed.
This also fixed the issue of not being able to select the first server in the drop down
list of DMB for Android. Closes #12

Changed redirect handling to manual to observe where the error occures.
Login seems to work but logout happens on loading of overview page.